requirements.txt
openai
requests
psycopg2-binary 